# [DRAFT] [onert] Eliminate copies of subgraph call/return