--- a/runtime/onert/core/src/backend/controlflow/IOTensor.cc
+++ b/runtime/onert/core/src/backend/controlflow/IOTensor.cc
@@ -38,7 +38,7 @@ IOTensor::IOTensor(IPortableTensor *tensor)
 }
 */
 
-void IOTensor::setTensor(ITensor *tensor)
+void IOTensor::setTensor(IPortableTensor *tensor)
 {
   assert(tensor);
   assert(tensor != this);