… runtime/onert/core/src/compiler/Compiler.cc
… runtime/onert/core/src/compiler/ExecutorFactory.cc
… runtime/onert/core/src/compiler/LoweredGraph.cc
… runtime/onert/core/src/compiler/StaticShapeInference.cc
… runtime/onert/core/src/compiler/pass/PermutationInsertionPass.cc
… runtime/onert/core/src/exec/DataflowExecutor.cc
… runtime/onert/core/src/exec/DataflowExecutor.h
… runtime/onert/core/src/exec/ExecutorBase.cc
… runtime/onert/core/src/exec/ExecutorBase.h
… runtime/onert/core/src/exec/LinearExecutor.cc
… runtime/onert/core/src/exec/LinearExecutor.h
… runtime/onert/core/src/exec/ParallelExecutor.cc
… runtime/onert/core/src/exec/ParallelExecutor.h