- Do not handle a primary graph specially
    - Make model input/outputs `IOTensor` rather than `UserTensor`
    - Also a preparation for NNAPI 1.3 support (Any subgraph can be primary)

ONE-DCO-1.0-Signed-off-by: Hanjoung Lee <hanjoung.lee@samsung.com>