@wateret 2020-10-16
TODO

- [ ] Change `ITensor` to `IPortableTensor` (If/While layer, IOTensor, ...)
- [ ] Remove `permute_subg_output_to_op_output` and related stuff in `IfLayer::run`
- [ ] Remove empty output tensor calls (`cond_exec->execute(_output_tensors, {})`)