@wateret 2020-10-16
Tests were passed before rebase.

It turns out that it fails since `getShape` has been devirtualized since #4614 . Indirecting tensor(`IOTensor`) does not work. 😭