@wateret 2020-10-27
Tested large-size If and While models with `nnfw_api_gtest`. (This is not going to be merged, just to test how much performance improvement)

<details>
  <summary>TEMPORARY PATCH TO TEST</summary>

```diff
--- a/tests/nnfw_api/src/GenModelTest.h
+++ b/tests/nnfw_api/src/GenModelTest.h
@@ -20,6 +20,7 @@
 #include <fstream>
 #include <string>
 #include <unordered_map>
+#include <chrono>
 
 #include "CircleGen.h"
 #include "fixtures.h"
@@ -356,7 +357,11 @@ protected:
           continue;
         }
 
+
+        std::chrono::steady_clock::time_point begin = std::chrono::steady_clock::now();
         NNFW_ENSURE_SUCCESS(nnfw_run(_so.session));
+        std::chrono::steady_clock::time_point end = std::chrono::steady_clock::now();
+        std::cout << "nnfw_run elapsed time : " << std::chrono::duration_cast<std::chrono::milliseconds>(end - begin).count() << "ms" << std::endl;
 
         ASSERT_EQ(_so.outputs.size(), ref_outputs.size());
         for (uint32_t i = 0; i < _so.outputs.size(); i++)
diff --git a/tests/nnfw_api/src/one_op_tests/If.cc b/tests/nnfw_api/src/one_op_tests/If.cc
index 4ec29422..c42ac52d 100644
--- a/tests/nnfw_api/src/one_op_tests/If.cc
+++ b/tests/nnfw_api/src/one_op_tests/If.cc
@@ -75,6 +75,73 @@ TEST_F(GenModelTest, OneOp_If)
   SUCCEED();
 }
 
+TEST_F(GenModelTest, OneOp_If_Large)
+{
+  // The model looks just like the below pseudocode
+  //
+  // function model(x, a, b)
+  // {
+  //   if (x)
+  //     return a;
+  //   else
+  //     return b;
+  // }
+
+  CircleGen cgen;
+
+  const int kElems = 32 * 1024 * 1024;
+  const std::vector<int32_t> shape{kElems};
+
+  // constant buffers
+  std::vector<float> then_data(kElems, -100);
+  uint32_t then_buf = cgen.addBuffer(then_data);
+  std::vector<float> else_data(kElems, 100);
+  uint32_t else_buf = cgen.addBuffer(else_data);
+
+  // primary subgraph
+  {
+    int x = cgen.addTensor({{1}, circle::TensorType_BOOL});
+    int a = cgen.addTensor({shape, circle::TensorType_FLOAT32, then_buf});
+    int b = cgen.addTensor({shape, circle::TensorType_FLOAT32, else_buf});
+    int ret = cgen.addTensor({shape, circle::TensorType_FLOAT32});
+    cgen.addOperatorIf({{x, a, b}, {ret}}, 1, 2);
+    cgen.setInputsAndOutputs({x}, {ret});
+  }
+
+  // then subgraph
+  {
+    cgen.nextSubgraph();
+    int a = cgen.addTensor({shape, circle::TensorType_FLOAT32});
+    int b = cgen.addTensor({shape, circle::TensorType_FLOAT32});
+    cgen.setInputsAndOutputs({a, b}, {a});
+  }
+
+  // else subgraph
+  {
+    cgen.nextSubgraph();
+    int a = cgen.addTensor({shape, circle::TensorType_FLOAT32});
+    int b = cgen.addTensor({shape, circle::TensorType_FLOAT32});
+    cgen.setInputsAndOutputs({a, b}, {b});
+  }
+
+  _context = std::make_unique<GenModelTestContext>(cgen.finish());
+  {
+    TestCaseData tcd;
+    tcd.addInput<bool>({true});
+    tcd.addOutput(then_data);
+    _context->addTestCase(tcd);
+  }
+  {
+    TestCaseData tcd;
+    tcd.addInput<bool>({false});
+    tcd.addOutput(else_data);
+    _context->addTestCase(tcd);
+  }
+  _context->setBackends({"cpu"});
+
+  SUCCEED();
+}
+
 class IfWrongSubgraphIndex : public GenModelTest,
                              public ::testing::WithParamInterface<std::pair<int, int>>
 {
diff --git a/tests/nnfw_api/src/one_op_tests/While.cc b/tests/nnfw_api/src/one_op_tests/While.cc
index 8b909c29..e138e3a0 100644
--- a/tests/nnfw_api/src/one_op_tests/While.cc
+++ b/tests/nnfw_api/src/one_op_tests/While.cc
@@ -75,6 +75,73 @@ TEST_F(GenModelTest, OneOp_While)
   SUCCEED();
 }
 
+TEST_F(GenModelTest, OneOp_While_Large)
+{
+  // The model looks just like the below pseudocode
+  //
+  // function model(x, data)
+  // {
+  //   // `data` does not do anything but passed to while's cond and body subgraphs
+  //   // to measure copy overhead between subgraphs
+  //   while (x < 100.0)
+  //   {
+  //     x = x + 1.0;
+  //   }
+  //   return (x, data)
+  // }
+
+  const int kElems = 1 * 1024 * 1024;
+  const std::vector<int32_t> shape{kElems};
+
+  CircleGen cgen;
+  uint32_t incr_buf = cgen.addBuffer(std::vector<float>{1});
+  uint32_t incr_data_buf = cgen.addBuffer(std::vector<float>(kElems, 1));
+  uint32_t end_buf = cgen.addBuffer(std::vector<float>{100});
+
+  // primary subgraph
+  {
+    int x_in = cgen.addTensor({{1}, circle::TensorType_FLOAT32});
+    int d_in = cgen.addTensor({shape, circle::TensorType_FLOAT32});
+    int x_out = cgen.addTensor({{1}, circle::TensorType_FLOAT32});
+    int d_out = cgen.addTensor({shape, circle::TensorType_FLOAT32});
+    cgen.addOperatorWhile({{x_in, d_in}, {x_out, d_out}}, 1, 2);
+    cgen.setInputsAndOutputs({x_in, d_in}, {x_out, d_out});
+  }
+
+  // cond subgraph
+  {
+    cgen.nextSubgraph();
+    int x = cgen.addTensor({{1}, circle::TensorType_FLOAT32});
+    int d = cgen.addTensor({shape, circle::TensorType_FLOAT32});
+    int end = cgen.addTensor({{1}, circle::TensorType_FLOAT32, end_buf});
+    int result = cgen.addTensor({{1}, circle::TensorType_BOOL});
+    cgen.addOperatorLess({{x, end}, {result}});
+    cgen.setInputsAndOutputs({x, d}, {result});
+  }
+
+  // body subgraph
+  {
+    cgen.nextSubgraph();
+    int x_in = cgen.addTensor({{1}, circle::TensorType_FLOAT32});
+    int incr = cgen.addTensor({{1}, circle::TensorType_FLOAT32, incr_buf});
+    int x_out = cgen.addTensor({{1}, circle::TensorType_FLOAT32});
+    int d_in = cgen.addTensor({shape, circle::TensorType_FLOAT32});
+    int incr_d = cgen.addTensor({shape, circle::TensorType_FLOAT32, incr_data_buf});
+    int d_out = cgen.addTensor({shape, circle::TensorType_FLOAT32});
+    cgen.addOperatorAdd({{x_in, incr}, {x_out}}, circle::ActivationFunctionType_NONE);
+    cgen.addOperatorAdd({{d_in, incr_d}, {d_out}}, circle::ActivationFunctionType_NONE);
+    cgen.setInputsAndOutputs({x_in, d_in}, {x_out, d_out});
+  }
+
+  _context = std::make_unique<GenModelTestContext>(cgen.finish());
+  std::vector<float> tc_data_in(kElems, 9);
+  std::vector<float> tc_data_out(kElems, 109);
+  _context->addTestCase(uniformTCD<float>({{0}, tc_data_in}, {{100}, tc_data_out}));
+  _context->setBackends({"cpu"});
+
+  SUCCEED();
+}
+
 TEST_F(GenModelTest, OneOp_While_TwoInputs)
 {
   // The model looks just like the below pseudocode

```
</details>

### Test cases

- Almost no other operations other than If/While
- If/While inputs/outputs are huge (32MB for If, 1MB for While)
    - must carry a lot of data while processing If/While op
    - so we can see the impact of copies when subgraph call/return
- Many iterations for While

### Result

If test : 197ms → 61ms
While test : 1391ms → 557ms

<details>
  <summary>TEST COMMAND AND OUTPUT</summary>

Master

```
$ ONERT_LOG_ENABLE= GRAPH_DOT_DUMP= Product/out/unittest_standalone/nnfw_api_gtest --gtest_filter="GenModelTest.OneOp_While_Large:GenModelTest.OneOp_If_Large"
Note: Google Test filter = GenModelTest.OneOp_While_Large:GenModelTest.OneOp_If_Large
[==========] Running 2 tests from 1 test case.
[----------] Global test environment set-up.
[----------] 2 tests from GenModelTest
[ RUN      ] GenModelTest.OneOp_If_Large
nnfw_run 210ms
nnfw_run 184ms
[       OK ] GenModelTest.OneOp_If_Large (15079 ms)
[ RUN      ] GenModelTest.OneOp_While_Large
nnfw_run 1391ms
[       OK ] GenModelTest.OneOp_While_Large (1746 ms)
[----------] 2 tests from GenModelTest (16825 ms total)

[----------] Global test environment tear-down
[==========] 2 tests from 1 test case ran. (16825 ms total)
[  PASSED  ] 2 tests.
```

Master + this PR

```
$ ONERT_LOG_ENABLE= GRAPH_DOT_DUMP= Product/out/unittest_standalone/nnfw_api_gtest --gtest_filter="GenModelTest.OneOp_While_Large:GenModelTest.OneOp_If_Large"
Note: Google Test filter = GenModelTest.OneOp_While_Large:GenModelTest.OneOp_If_Large
[==========] Running 2 tests from 1 test case.
[----------] Global test environment set-up.
[----------] 2 tests from GenModelTest
[ RUN      ] GenModelTest.OneOp_If_Large
nnfw_run 63ms
nnfw_run 58ms
[       OK ] GenModelTest.OneOp_If_Large (5226 ms)
[ RUN      ] GenModelTest.OneOp_While_Large
nnfw_run 557ms
[       OK ] GenModelTest.OneOp_While_Large (722 ms)
[----------] 2 tests from GenModelTest (5949 ms total)

[----------] Global test environment tear-down
[==========] 2 tests from 1 test case ran. (5949 ms total)
[  PASSED  ] 2 tests.
```

</details>